brigand::compose moved in types/compose.hpp
