+++ brigand/sequences/set.hpp
@@ -8,6 +8,7 @@
 #include <type_traits>
 #include <brigand/types/type.hpp>
 #include <brigand/types/bool.hpp>
+#include <brigand/types/compose.hpp>
 #include <brigand/algorithms/apply.hpp>
 #include <brigand/sequences/list.hpp>
 #include <brigand/sequences/erase.hpp>
@@ -36,7 +37,7 @@ namespace detail
 
         template<class K>
         static apply<typename erase_if_impl<list<T...>, unary_compose<std::is_same, K>::template type, list<>>::type, detail::set_impl> erase(type_<K>);
-    
+
     private:
         template<class K>
         static set_impl<T..., K> insert_impl(false_);
